requirements.txt
fastapi==0.115.9
uvicorn==0.27.1
sentence-transformers==2.5.1
qdrant-client==1.7.0
google-generativeai==0.3.1
pydantic==2.11.4 